refactor: settings main page
