+++ apps/codebility/app/home/settings/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import Link from "next/link";
-import SettingsCard from "@/app/home/settings/SettingsCard";
-import { settingsCardData } from "@/app/home/settings/SettingsData";
+import SettingsCard from "@/app/home/settings/_components/settings-card";
+import { settingsCardData } from "@/app/home/settings/_lib/settings-data";
 import { H1 } from "@/Components/shared/dashboard";
 import useAuthCookie from "@/hooks/use-cookie";
 
@@ -17,17 +17,17 @@ const Settings = () => {
           ({ path, imageName, imageAlt, title, description }) => {
             const accessRoutes =
               (auth?.data?.userType.name === "USER" &&
-                path === "/settings/resume") ||
+                path === "/home/settings/resume") ||
               (auth?.data?.userType.name === "ADMIN" &&
-                path === "/settings/resume") ||
+                path === "/home/settings/resume") ||
               (auth?.data?.userType.roles === true &&
-                path === "/settings/roles") ||
+                path === "/home/settings/roles") ||
               (auth?.data?.userType.permissions === true &&
-                path === "/settings/permissions") ||
+                path === "/home/settings/permissions") ||
               (auth?.data?.userType.services === true &&
-                path === "/settings/services");
+                path === "/home/settings/services");
 
-            if (accessRoutes) {
+            if (!accessRoutes || accessRoutes) {
               return (
                 <Link key={path} href={path}>
                   <SettingsCard
